Passer a du mobile

+++ services/storage.ts
@@ -1,36 +1,63 @@
 import { AgendaEvent } from "@/types/agenda";
 import { Note } from "@/types/note";
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import * as SecureStore from 'expo-secure-store';
+import { Platform } from "react-native";
+const isWeb = Platform.OS === 'web';
 
 const tokenName = "Token";
 const userName = "User";
+const passwordName = "Password";
 const agendaName = "Agenda";
 const notesName = "Notes";
 
-export function getToken() {
-  return localStorage.getItem(tokenName);
+const getStorageItem = async (key: string): Promise<string | null> => {
+  if (isWeb) {
+    return localStorage.getItem(key);
+  }
+  return await AsyncStorage.getItem(key);
+};
+
+const setStorageItem = async (key: string, value: string) => {
+  if (isWeb) {
+    localStorage.setItem(key, value);
+  } else {
+    await AsyncStorage.setItem(key, value);
+  }
+};
+export async function getToken(): Promise<string | null> {
+  return await getStorageItem(tokenName);
 }
-export function setToken(value: string) {
-  return localStorage.setItem(tokenName, value);
+export async function setToken(value: string) {
+  await setStorageItem(tokenName, value);
 }
 
-export function getId() {
-  return localStorage.getItem(userName);
+export async function getId(): Promise<string | null> {
+  return await getStorageItem(userName);
 }
-export function setId(value: string) {
-  return localStorage.setItem(userName, value);
+export async function setId(value: string) {
+  await setStorageItem(userName, value);
 }
 
-export const saveAgendaToCache = (parsedEvents: AgendaEvent[]) => {
+export async function getPasswordStorage(): Promise<string | null> {
+  return await getStorageItem(passwordName);
+}
+export async function setPasswordStorage(value: string) {
+  await setStorageItem(passwordName, value);
+}
+
+export async function saveAgendaToCache(parsedEvents: AgendaEvent[]) {
+
   try {
-    localStorage.setItem(agendaName, JSON.stringify(parsedEvents));
-    localStorage.setItem(agendaName+"Date", Date.now().toString()); // Optionnel : pour savoir quand on a mis en cache
+    setStorageItem(agendaName, JSON.stringify(parsedEvents));
+    setStorageItem(agendaName + "Date", Date.now().toString());
   } catch (e) {
     console.error("Quota localStorage dépassé", e);
   }
 };
 
-export const loadAgendaFromCache = (): AgendaEvent[] | null => {
-  const cachedString = localStorage.getItem(agendaName);
+export async function loadAgendaFromCache(): Promise<AgendaEvent[] | null> {
+  const cachedString = await getStorageItem(agendaName);
 
   if (!cachedString) return null;
 
@@ -45,37 +72,68 @@ export const loadAgendaFromCache = (): AgendaEvent[] | null => {
   }));
 
   return fixedEvents;
+
 };
 
-export const saveNotesToCache = (notes: Note[]) => {
+export async function saveNotesToCache(notes: Note[]) {
   try {
-    localStorage.setItem(notesName, JSON.stringify(notes));
-    localStorage.setItem(notesName+"Date", Date.now().toString()); // Optionnel : pour savoir quand on a mis en cache
+    await setStorageItem(notesName, JSON.stringify(notes));
+    await setStorageItem(notesName + "Date", Date.now().toString());
   } catch (e) {
     console.error("Quota localStorage dépassé", e);
   }
 };
 
-export const loadNotesFromCache = (): Note[] | null => {
-  const cachedString = localStorage.getItem(notesName);
+export async function loadNotesFromCache(): Promise<Note[] | null> {
+  const cachedString = await getStorageItem(notesName);
 
   if (!cachedString) return null;
 
   // On transforme le texte en objet JSON
   const rawNotes = JSON.parse(cachedString);
   return rawNotes;
+
 };
 
+// Remove
+async function removeStorageItem(key: string) {
+  if (isWeb) {
+    localStorage.removeItem(key);
+  } else {
+    await AsyncStorage.removeItem(key);
+  }
+}
 
-export const clearAgendaFromStorage = () => {
-  localStorage.removeItem(agendaName);
-  localStorage.removeItem(agendaName+"Date");
-};
+export async function clearAgendaFromStorage() {
+  await Promise.all([
+    removeStorageItem(agendaName),
+    removeStorageItem(agendaName + "Date")
+  ]);
+}
 
-export const clearNotesFromStorage = () => {
-  localStorage.removeItem(notesName);
-  localStorage.removeItem(notesName+"Date");
-};
-export const clearAllStorage = () => {
-  localStorage.clear();
+export async function clearNotesFromStorage() {
+  await Promise.all([
+    removeStorageItem(notesName),
+    removeStorageItem(notesName + "Date")
+  ]);
+}
+
+export async function clearAppCache() {
+  await Promise.all([
+    clearAgendaFromStorage(),
+    clearNotesFromStorage()
+  ]);
+}
+
+export async function clearAllStorage() {
+  if (isWeb) {
+    localStorage.clear();
+    return;
+  }
+
+  await AsyncStorage.clear();
+  await Promise.all([
+    SecureStore.deleteItemAsync(tokenName),
+    SecureStore.deleteItemAsync(userName)
+  ]);
 }
